+++ completion/git-machete.completion.bash
@@ -3,13 +3,14 @@
 _git_machete() {
     cmds="add anno d delete-unmanaged diff discover e edit file fork-point g go help l list log reapply show slide-out s status traverse update"
     help_topics="$cmds format hooks"
-    categories="managed slidable slidable-after unmanaged"
+    categories="managed slidable slidable-after unmanaged with-overridden-fork-point"
     directions="down first last next prev root up"
 
     common_opts="--debug -h --help -v --verbose --version"
     add_opts="-o --onto="
     diff_opts="-s --stat"
     discover_opts="-l --list-commits -r --roots="
+    fork_point_opts="--override-to= --unset-override"
     reapply_opts="-f --fork-point="
     slide_out_opts="-d --down-fork-point="
     status_opts="-l --list-commits"
@@ -17,7 +18,7 @@ _git_machete() {
     update_opts="-f --fork-point="
 
     case "$cur" in
-        --down-fork-point=*|--fork-point=*) __gitcomp "$(__git_refs)" "" "${cur##--*=}" ;;
+        --down-fork-point=*|--fork-point=*|--override-to=*) __gitcomp "$(__git_refs)" "" "${cur##--*=}" ;;
         --onto=*) __gitcomp_nl "$(git machete list managed)" "" "${cur##--onto=}" ;;
         --roots=*) __gitcomp "$(__git_heads)" "" "${cur##--roots=}" ;;
         -*)
@@ -25,6 +26,7 @@ _git_machete() {
                 add) __gitcomp "$common_opts $add_opts" ;;
                 d|diff) __gitcomp "$common_opts $diff_opts" ;;
                 discover) __gitcomp "$common_opts $discover_opts" ;;
+                fork-point) __gitcomp "$common_opts $fork_point_opts" ;;
                 reapply) __gitcomp "$common_opts $reapply_opts" ;;
                 slide-out) __gitcomp "$common_opts $slide_out_opts" ;;
                 s|status) __gitcomp "$common_opts $status_opts" ;;
@@ -38,13 +40,14 @@ _git_machete() {
              else
                 prev="${COMP_WORDS[COMP_CWORD-1]}"
                 case "$prev" in
-                    -d|--down-fork-point|-f|--fork-point) __gitcomp "$(__git_refs)" ;;
+                    -d|--down-fork-point|-f|--fork-point|--override-to) __gitcomp "$(__git_refs)" ;;
                     # TODO #25: We don't complete --help since it's going to be captured by git anyway
                     # (and results in redirection to yet non-existent man for `git-machete`).
                     -h) __gitcomp "$help_topics" ;;
                     -o|--onto) __gitcomp_nl "$(git machete list managed)" ;;
                     # TODO complete the comma-separated list of roots
                     -r|--roots) __gitcomp "$(__git_heads)" ;;
+                    --unset-override) __gitcomp_nl "$(git machete list with-overridden-fork-point)" ;;
                     *)
                         case "${COMP_WORDS[2]}" in
                             add) __gitcomp_nl "$(git machete list unmanaged)" ;;